Added deferred rendering

+++ src/lib/vengine/shaders/include/structs.glsl
@@ -19,7 +19,7 @@ struct SceneData {
     mat4 projectionInverse;
     vec4 exposure;      /* R = exposure, G = environment map intensity, B = lens radius , A = focal distance */
     vec4 background;    /* RGB = background color, A = environment type */
-    ivec4 volumes;      /* R = material id of camera volume, GBA = unused */
+    vec4 volumes;      /* R = material id of camera volume, G = near plane, B = far plane, A = unused */
 };
 
 /* InstanceData struct. A mirror of the CPU struct */
@@ -40,12 +40,12 @@ struct InstanceData {
 /* Material struct. A mirror of the CPU struct */
 struct MaterialData
 {
-    vec4 albedo;                /* RGB: albedo, A: unused */
+    vec4 albedo;                /* RGB: albedo, A: alpha */
     vec4 metallicRoughnessAO;   /* R: metallic, G: roughness, B: AO, A = is transparent */
     vec4 emissive;              /* RGB: emissive color, A = emissive intensity */
     uvec4 gTexturesIndices1;    /* R: albedo texture index, G: metallic texture index, B: roughness texture index, A: AO texture index */   
     uvec4 gTexturesIndices2;    /* R: emissive texture index, G: normal texture index, B: BRDF LUT texture index, A: unused */
-    vec4 uvTiling;              /* R: u tiling, G: v tiling, B: unused, A: unused */
+    vec4 uvTiling;              /* R: u tiling, G: v tiling, B: material type, A: unused */
     
     uvec4 padding1;
     uvec4 padding2;
